refactor: TokenValidationResult를 model 패키지로 이동
