+++ backend/codelight/src/main/java/com/project/codelight/auth/service/AuthService.java
@@ -7,7 +7,7 @@ import com.project.codelight.auth.dto.response.ReissueTokenResponse;
 import com.project.codelight.auth.repository.AuthRepository;
 import com.project.codelight.auth.repository.RefreshTokenRepository;
 import com.project.codelight.auth.repository.TokenBlackListRepository;
-import com.project.codelight.auth.service.dto.TokenValidationResult;
+import com.project.codelight.auth.service.model.TokenValidationResult;
 import com.project.codelight.auth.util.TokenUtils;
 import com.project.codelight.global.exception.CodeLightException;
 import com.project.codelight.global.exception.ExceptionCodeType;
